Rebuild with GCC rebased against 5f9d38fd8; remove 32-bit binaries.
There have been some changes since then which affect us; in particular,
enabling relro breaks elfhack.  Since the goal is to be as close to the
existing toolchain as possible, let's use the same commit.

Rebuilt for Linux and Mac x86_64; removed the i686 Linux build instead,
because I don't think anyone uses it, and it may not even work.